initial pass at refactoring prolly.Range
