+++ go/store/prolly/map.go
@@ -17,8 +17,6 @@ package prolly
 import (
 	"context"
 	"fmt"
-	"io"
-	"sort"
 
 	"github.com/dolthub/dolt/go/store/hash"
 	"github.com/dolthub/dolt/go/store/types"
@@ -166,11 +164,7 @@ func (m Map) Last(ctx context.Context) (key, value val.Tuple, err error) {
 
 // IterAll returns a MutableMapRangeIter that iterates over the entire Map.
 func (m Map) IterAll(ctx context.Context) (MapRangeIter, error) {
-	rng := Range{
-		Start:   RangeCut{Unbound: true},
-		Stop:    RangeCut{Unbound: true},
-		KeyDesc: m.keyDesc,
-	}
+	rng := Range{Start: nil, Stop: nil, KeyDesc: m.keyDesc}
 	return m.IterRange(ctx, rng)
 }
 
@@ -186,21 +180,21 @@ func (m Map) iterFromRange(ctx context.Context, rng Range) (*prollyRangeIter, er
 		stop  *nodeCursor
 	)
 
-	startSearch := m.rangeStartSearchFn(rng)
-	if rng.Start.Unbound {
+	startSearch := rangeStartSearchFn(rng)
+	if rng.Start == nil {
 		start, err = newCursorAtStart(ctx, m.ns, m.root)
 	} else {
-		start, err = newCursorAtTuple(ctx, m.ns, m.root, rng.Start.Key, startSearch)
+		start, err = newCursorFromSearchFn(ctx, m.ns, m.root, startSearch)
 	}
 	if err != nil {
 		return nil, err
 	}
 
-	stopSearch := m.rangeStopSearchFn(rng)
-	if rng.Stop.Unbound {
+	stopSearch := rangeStopSearchFn(rng)
+	if rng.Stop == nil {
 		stop, err = newCursorPastEnd(ctx, m.ns, m.root)
 	} else {
-		stop, err = newCursorAtTuple(ctx, m.ns, m.root, rng.Stop.Key, stopSearch)
+		stop, err = newCursorFromSearchFn(ctx, m.ns, m.root, stopSearch)
 	}
 	if err != nil {
 		return nil, err
@@ -214,42 +208,6 @@ func (m Map) iterFromRange(ctx context.Context, rng Range) (*prollyRangeIter, er
 		curr: start,
 		stop: stop,
 	}, nil
-}
-
-func (m Map) rangeStartSearchFn(rng Range) searchFn {
-	// todo(andy): inline sort.Search()
-	return func(query nodeItem, nd Node) int {
-		return sort.Search(int(nd.count), func(i int) bool {
-			q := val.Tuple(query)
-			t := val.Tuple(nd.getKey(i))
-
-			// compare using the range's tuple descriptor.
-			cmp := rng.KeyDesc.Compare(q, t)
-			if rng.Start.Inclusive {
-				return cmp <= 0
-			} else {
-				return cmp < 0
-			}
-		})
-	}
-}
-
-func (m Map) rangeStopSearchFn(rng Range) searchFn {
-	// todo(andy): inline sort.Search()
-	return func(query nodeItem, nd Node) int {
-		return sort.Search(int(nd.count), func(i int) bool {
-			q := val.Tuple(query)
-			t := val.Tuple(nd.getKey(i))
-
-			// compare using the range's tuple descriptor.
-			cmp := rng.KeyDesc.Compare(q, t)
-			if rng.Stop.Inclusive {
-				return cmp < 0
-			} else {
-				return cmp <= 0
-			}
-		})
-	}
 }
 
 // searchNode returns the smallest index where nd[i] >= query
@@ -274,7 +232,7 @@ func (m Map) searchNode(query nodeItem, nd Node) int {
 	return i
 }
 
-var _ searchFn = Map{}.searchNode
+var _ itemSearchFn = Map{}.searchNode
 
 // compareItems is a compareFn.
 func (m Map) compareItems(left, right nodeItem) int {
@@ -284,57 +242,4 @@ func (m Map) compareItems(left, right nodeItem) int {
 
 func (m Map) compareKeys(left, right val.Tuple) int {
 	return int(m.keyDesc.Compare(left, right))
-}
-
-type prollyRangeIter struct {
-	// current tuple location
-	curr *nodeCursor
-	// non-inclusive range stop
-	stop *nodeCursor
-}
-
-var _ rangeIter = &prollyRangeIter{}
-var _ MapRangeIter = &prollyRangeIter{}
-
-func (it *prollyRangeIter) Next(ctx context.Context) (key, value val.Tuple, err error) {
-	if it.curr == nil {
-		return nil, nil, io.EOF
-	}
-
-	key = it.curr.nd.keys.GetSlice(it.curr.idx)
-	value = it.curr.nd.values.GetSlice(it.curr.idx)
-
-	_, err = it.curr.advance(ctx)
-	if err != nil {
-		return nil, nil, err
-	}
-	if it.curr.compare(it.stop) >= 0 {
-		// past the end of the range
-		it.curr = nil
-	}
-
-	return
-}
-
-func (it *prollyRangeIter) current() (key, value val.Tuple) {
-	// |it.curr| is set to nil when its range is exhausted
-	if it.curr != nil && it.curr.valid() {
-		key = it.curr.nd.keys.GetSlice(it.curr.idx)
-		value = it.curr.nd.values.GetSlice(it.curr.idx)
-	}
-	return
-}
-
-func (it *prollyRangeIter) iterate(ctx context.Context) (err error) {
-	_, err = it.curr.advance(ctx)
-	if err != nil {
-		return err
-	}
-
-	if it.curr.compare(it.stop) >= 0 {
-		// past the end of the range
-		it.curr = nil
-	}
-
-	return
 }
