+++ go/store/val/codec.go
@@ -409,61 +409,6 @@ func writeRaw(buf, val []byte) {
 	copy(buf, val)
 }
 
-func compare(typ Type, left, right []byte) int {
-	// order NULLs last
-	if left == nil {
-		if right == nil {
-			return 0
-		} else {
-			return 1
-		}
-	} else if right == nil {
-		if left == nil {
-			return 0
-		} else {
-			return -1
-		}
-	}
-
-	switch typ.Enc {
-	case Int8Enc:
-		return compareInt8(readInt8(left), readInt8(right))
-	case Uint8Enc:
-		return compareUint8(readUint8(left), readUint8(right))
-	case Int16Enc:
-		return compareInt16(readInt16(left), readInt16(right))
-	case Uint16Enc:
-		return compareUint16(readUint16(left), readUint16(right))
-	case Int32Enc:
-		return compareInt32(readInt32(left), readInt32(right))
-	case Uint32Enc:
-		return compareUint32(readUint32(left), readUint32(right))
-	case Int64Enc:
-		return compareInt64(readInt64(left), readInt64(right))
-	case Uint64Enc:
-		return compareUint64(readUint64(left), readUint64(right))
-	case Float32Enc:
-		return compareFloat32(readFloat32(left), readFloat32(right))
-	case Float64Enc:
-		return compareFloat64(readFloat64(left), readFloat64(right))
-	case YearEnc:
-		return compareInt16(readInt16(left), readInt16(right))
-	case DateEnc, DatetimeEnc, TimestampEnc:
-		return compareTimestamp(readTimestamp(left), readTimestamp(right))
-	case TimeEnc:
-		panic("unimplemented")
-	case DecimalEnc:
-		// todo(andy): temporary Decimal implementation
-		fallthrough
-	case StringEnc:
-		return compareString(readString(left), readString(right))
-	case ByteStringEnc:
-		return compareByteString(readByteString(left), readByteString(right))
-	default:
-		panic("unknown encoding")
-	}
-}
-
 func expectSize(buf []byte, sz ByteSize) {
 	if ByteSize(len(buf)) != sz {
 		panic("byte slice is not of expected size")
